relocate metadata encryption - part 4
